src/python/workshop/requirements.txt
aiosqlite>=0.20.0, <1.0.0
httpx>=0.27.2, <0.28.0
aiohttp>=3.11.11, <4.0.0
python_dotenv>=1.0.1, <2.0.0
azure-identity>=1.19.0, <2.0.0
azure-ai-projects==1.0.0b5
pandas>=2.2.3, <3.0.0
pydantic==2.10.1
pillow>=11.1.0, <12.0.0